renamed user model transfered function to controllers
