+++ backend/controllers/project.controller.ts
@@ -1,0 +1,1 @@
+// Placeholder text
